requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
requests>=2.25.0
scipy>=1.7.0
scikit-learn>=1.0.0
werkzeug>=2.0.0
gunicorn>=20.0.0
python-dotenv>=0.19.0
imageio>=2.9.0
matplotlib>=3.3.0